refactor: implement logic for internal list-view lnks, action-bar titles and context passing
