+++ app/application/check-platform/check-platform-page.ts
@@ -1,10 +1,11 @@
 // >> application-import-ts
 import * as application from "tns-core-modules/application";
 // << application-import-ts
-import { EventData } from "tns-core-modules/data/observable";
+import { EventData, Observable } from "tns-core-modules/data/observable";
 import { Color } from "tns-core-modules/color";
 import { Label } from "tns-core-modules/ui/label";
 import { GridLayout } from "tns-core-modules/ui/layouts/grid-layout";
+import { Page } from "tns-core-modules/ui/page";
 
 export function onGridLoaded(args: EventData) {
     const grid = <GridLayout>args.object;
@@ -22,6 +23,7 @@ export function onGridLoaded(args: EventData) {
         lbl.fontSize = 26;
         lbl.text = "Android Applicaiton";
         lbl.textAlignment = "center";
+        lbl.verticalAlignment = "middle";
         // << (hide)
     } else if (application.ios) {
         console.log("We are running on iOS device");
@@ -37,4 +39,18 @@ export function onGridLoaded(args: EventData) {
         // << (hide)
     }
     // << application-platform-ts
+}
+
+export function onNavigatingTo(args) {
+    let page = <Page>args.object;
+    page.actionBar.title = "";
+}
+
+export function onNavigatedTo(args) {
+    let page = <Page>args.object;
+
+    let vm = new Observable();
+    vm.set("actionBarTitle", args.context.actionBarTitle);
+
+    page.bindingContext = vm;
 }
