+++ app/application/check-platform/check-platform-page.xml
@@ -1,4 +1,8 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" loaded="onLoaded" navigatedTo="onNavigatedTo">
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" navigatedTo="onNavigatedTo">
+    <Page.actionBar>
+		<ActionBar title="{{ actionBarTitle }}" icon="" class="action-bar">
+		</ActionBar>
+	</Page.actionBar>
     <GridLayout rows="*" columns="*" loaded="onGridLoaded">
         <Label id="lbl" textWrap="true" />
     </GridLayout>
