+++ package.json
@@ -9,7 +9,7 @@
       "version": "3.4.1-2018-02-27-03"
     },
     "tns-ios": {
-      "version": "next"
+      "version": "4.0.0-2018-02-26-01"
     }
   },
   "dependencies": {
